requirements.txt
asgiref==3.7.2
Django
sqlparse==0.4.4
django-schema-graph